# storage/docs/openapi.yaml
openapi: 3.0.3
info:
  description: "This is the Collectio Server API"
  version: 1.0.0
  title: Collectio
  termsOfService: https://collectioapp.com/
  contact:
    email: support@collectioapp.com
  license:
    name: MIT
    url: https://mit-license.org
servers:
  - url: https://collectioapp.com/api/v1
  - url: https://beta.collectioapp.com/api/v1
security:
  - basicAuth: []
  - apiKey: []
...
paths:
...
  /user:
    get:
      tags:
        - user
      summary: Get user
      description: ""
      operationId: getUser
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/responses/User"
    post:
      tags:
        - user
      summary: Edit user
      description: ""
      operationId: editUser
      requestBody:
        description: ""
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/responses/User"
  /user/image:
    delete:
      tags:
        - user
      summary: Delete user image
      description: ""
      operationId: deleteUserImage
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/responses/User"
  /user/api-keys:
    get:
      tags:
        - apiKeys
      summary: Get api keys
      description: ""
      operationId: listApiKeys
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: api token
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ListApiKeys"
    post:
      tags:
        - apiKeys
      summary: Edit apikey
      description: ""
      operationId: editApikey
      requestBody:
        description: ""
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Apikey"
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: api token
          content:
            application/json:
              schema:
                $ref: "#/components/responses/EditApikey"
    delete:
      tags:
        - apiKeys
      summary: Delete apikey
      description: ""
      operationId: deleteApikey
      parameters:
        - name: Token
          in: path
          description: Token to delete
          required: true
          explode: true
          schema:
            type: string
      responses:
        "500":
          description: Internal Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/InternalError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ForbiddenBasicAuth"
        "200":
          description: api token
          content:
            application/json:
              schema:
                $ref: "#/components/responses/DeleteApikey"
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Token
  responses:
    ForbiddenBasicAuth:
      $ref: "#/components/schemas/ResponseWS"
      # Code 403, Status: 'KO'
...
    InternalError:
      $ref: "#/components/schemas/ResponseWS"
      # Code 500, Status: 'KO'
...
  schemas:
    ResponseWS:
      type: object
      properties:
        status:
          type: string
          minLength: 2
          maxLength: 2
        code:
          type: integer
          format: int64
        message:
          type: string
        data:
          nullable: true
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Category"
            - type: array
              items:
                $ref: "#/components/schemas/Subcategory"
            - type: array
              items:
                $ref: "#/components/schemas/Collection"
            - type: array
              items:
                $ref: "#/components/schemas/Item"
            - type: array
              items:
                $ref: "#/components/schemas/ItemImage"
            - type: array
              items:
                $ref: "#/components/schemas/Apikey"
            - $ref: "#/components/schemas/User"
    Category:
      type: object
      properties:
        Id:
          type: integer
          format: int64
        Image:
          type: string
        Spanish:
          type: string
        English:
          type: string
        Catalan:
          type: string
        Basque:
          type: string
    Subcategory:
      type: object
      properties:
        Id:
          type: integer
          format: int64
        CategoryId:
          type: integer
          format: int64
        Image:
          type: string
        Spanish:
          type: string
        English:
          type: string
        Catalan:
          type: string
        Basque:
          type: string
    Collection:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        UserId:
          type: integer
          format: int64
        CategoryId:
          type: integer
          format: int64
        Name:
          type: string
        Description:
          type: string
        Private:
          type: boolean
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
        Image:
          type: string
        Items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Item:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        CollectionId:
          type: integer
          format: int64
        SubcategoryId:
          type: integer
          format: int64
        Name:
          type: string
        Description:
          type: string
        Private:
          type: boolean
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
        Images:
          type: array
          items:
            $ref: "#/components/schemas/ItemImage"
    ItemImage:
      type: object
      properties:
        ServerId:
          type: integer
          format: int64
        ItemId:
          type: integer
          format: int64
        Image:
          type: string
    User:
      type: object
      properties:
        Nickname:
          type: string
        Mail:
          type: string
        Image:
          type: string
        Apple:
          type: boolean
        Google:
          type: boolean
    Apikey:
      type: object
      properties:
        Token:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UsedAt:
          type: string
          format: date-time
        Device:
          type: string
        UserDeviceName:
          type: string